requirements.txt
replicate
streamlit==1.37.0
requests
streamlit-image-select
google-generativeai